Remove buffer-offset compensation in transport frames.
It was found that Ardour has a bug where it doesn't actually start
recording from software sources when the transport frame == 0.
Compensation for that had been added to Hydrogen.  This takes the
compensation out... but does it with a macro so that it can be
easily seen and re-enabled.

In future commits, this will become user-switchable.  However,
this code *works*, whereas I ran in to trouble with the more
configurable solution.

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@444 3b3fb362-3133-0410-aa15-cf69e0a59cb7