Ticket 31: Lead in lead/lag doesn't adjust properly.
User prschmitt on H2 forum reports that the lead in the lead/lag feature
doesn't adjust properly. Instead of adjusting the lead gradually, it
seemingly jumps from a small lead to a large lead, with nothing in
between.

audioEngine_updateNoteQueue() was checking to see if the offset put us
before the first tick in the process cycle (i.e. frame 0).  However,
the "first tick" that was being compared isn't actually the first
tick -- but had a buffer to prevent lead/lag from preceeding frame 0.
Therefore the check (and corrective action) has been removed.

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@467 3b3fb362-3133-0410-aa15-cf69e0a59cb7