Track lookahead globally in hydrogen.cpp, and compensate when recording realtime MIDI notes.
A forum user reported [1] that MIDI notes are *sounding* at the right time,
but being *recorded* about 1/16th late.  Sebastian Moors (smoors) said [2]
that he isolated to SVN rev 210, where lead-lag was introduced.  In
that commit, the value of m_nPatternTickPosition was adjusted to
include a lookahead factor.  This allowed notes with positive lead and
positive humanize (i.e. before the beat) to be correctly scheduled.
However, realtime notes were then recorded *with* the leadlag, putting
notes 5 ticks + 2000 frames late (typ. 8-11 ticks).  It was typically
not noticeable with quantize on -- but with quantize off, it was
always present.

This commit implements a global variable m_nLookaheadFrames.  When
doing realtime MIDI recording, it will compensate the "current" tick
to remove the lookahead.

This was done instead of changing the way that lookahead was
implemented.  Hydrogen has a lot of code that depends on the "current"
tick or the "current" pattern for proper scheduling.  The current
lookahead forces Hydrogen to live in the future... always a few
thousand frames ahead of the sound output.  To try and "correct" that
would prove fairly invasive to the application as a whole (as seen in
commits 527 and 528).

[1] http://www.hydrogen-music.org/forum/?action=show_thread&thread=882&fid=5&page=1
[2] e-mail to hydrogen-devel@hydrogen-music.org, 08SEP2008.

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@594 3b3fb362-3133-0410-aa15-cf69e0a59cb7