Replace jack_client_new() with jack_client_open().
The JACK driver was using the deprecated jack_client_new() function, and
manually trying to manage unique client names for Hydrogen.  This is
now done by the JACK server using the jack_client_open().  Multiple
tries to open the server have to do with the start/stop code in
the H2 AudioEngine.  This code also checks errors and to see if the
JACK server changed the client name on us.

Compatible with JACK >= 0.99.4

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@726 3b3fb362-3133-0410-aa15-cf69e0a59cb7