Fix memory leaks with QFileDialog.
When creating a QFileDialog, Qt creates at least two threads for the
dialog.  In Hydrogen, QFileDialogs were typically invoked like this:

   QFileDialog *fd = new QFileDialog(this);

That is, *fd is never deleted until the destructor for *this is called
(when Hydrogen is closed down).  Therefore, every time someone opened
a new file, two new threads would be created and persist.  With
several of these threads hanging around (on the order of 20), Hydrogen
bogs down -- and this is possibly causing xruns (see bug #85).

It is corrected like this:

   std::auto_ptr<QFileDialog> fd( new QFileDialog );

The pointer container fd is auto-deleted when it goes out of scope,
even if there is an exception.

This _possibly_ fixes bug #85.  It definately fixes the memory leak.

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@932 3b3fb362-3133-0410-aa15-cf69e0a59cb7