Drop audioEngine state when loading drum kit.
Loading a drumkit (Hydrogen::loadDrumkit()) is not RT safe because
some of the functions that it calls will load samples and allocate
memory for those samples.  Therefore, m_audioEngineState
(hydrogen.cpp) is dropped below STATE_READY so that the process()
callback will "short-circuit" while we wait on the drumkit to load.

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@1284 3b3fb362-3133-0410-aa15-cf69e0a59cb7