Load sane settings even if hydrogen.default.conf is missing.
If hydrogen starts up and can not find hydrogen.default.conf, or
$HOME/.hydrogen/hydrogen.conf, then it loads nonsense settings for
colors and window sizes and fonts.  This results in an unusable UI.

What's worse, when the user closes hydrogen, these nonsense settings
get SAVED to $HOME/.hydrogen/hydrogen.conf.  Then, the user will load
the nonsense settings every time they load hydrogen.

The only way to recover from this was to delete the user prefs file
and make sure that hydrogen.default.conf is restored.  However, it's
not obvious to any user that this is how to correct the problem.
Therefore, the defaults have been added to Preferences::Preferences()
to circumvent this situation.

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@1321 3b3fb362-3133-0410-aa15-cf69e0a59cb7