Adding Anaconda path in CMakeFile In case Anaconda is installed on the system it will easily mask some of the libraries and shared objects needed for compilation. The `find_` commands of CMAKE, e.g. `findZLIB`, seem to be agnostic of the lib path set up in the user home. As a result CMAKE finds e.g. `zlib` in /usr/lib but during the compilation gcc tries to link against the installation in $HOME/anaconda/lib causing the build to fail.
Via an `if` clause the library path of Anaconda will be included as soon as the string "anaconda" is present in the `PATH` environmental variable.

There is still a problem I do not quite know how to fix in a platform-independent way: one has to detect the actual path to the anaconda installation. For me it is `.anaconda3` but this is not the default and each user is able to choose her own one. Therefore it should be grepped from the `PATH` variable.