sampler: fixing holes in instrument layers In some drumkits, like the default GM ones, separate layers of an individual instrument are not perfectly aligned. E.g. for the "Hat Open" of the GM Kit the third layer ends at a velocity of 0.793478 and the fourth one starts at 0.797101. But since the selection of a sample for a specific note only checks whether its velocity is *within* each layer, the note can fall between two layers causing the sampler to skip it.
In this patch the sampling will work as it always does. But in the rare cases the note falls between two layers another for loop checks which layer is the closest one to the note and uses it for the sampling. This event should in general occur only occasionally so having a separate chunk so code handling it might be the most efficient way. In addition, the lookup of the nearest layer also respects instruments, which layers are not sorted in ascending order.