Filesystem : add methods to handle sample filepath
int get_basename_idx_within_drumkit(QString) returns the index of the
basename of the sample if the given path is under an existing drumkit
path, otherwise returns -1. The existence of the sample path is not checked !!

bool file_is_within_drumkit(QString) is a wrapper over the previous function

QString prepare_sample_path(fname) returns the basename if the given
path is under an existing drumkit, otherwise the given path.