Controlling multiple note properties + bug fixes in pattern editor
Lately I revisited some old songs for which I set the velocity
properties of the individual notes by hand. Now I want them all to
have a common offset and to leave the rest to the humanizer (apart
from some highlights). But changing all values separately is way too
tedious. Therefore, I implemented a way the properties of all notes in
one pattern (velocity, probability, pan etc.) can be set at once.

When holding the SHIFT key and moving the mouse wheel over the ruler
of the pattern editor all notes are affected by the changes. When
pressing the CTRL key too, the change is not +/- 0.05 but +/- 0.01 for
all of them. Holding the SHIFT key and dragging a property of one of
the notes all others are set to exactly the same value.

In addition, I fixed some bugs, which in the ruler:
- Pressing CTRL did only had an effect on the mouse wheel if no other
  modifier key was pressed
- Adjusting the **pan** and **lead and lag** was not working using the
  mouse wheel since the could did not change the proper
  variables. Instead of assigning the old and new values to *__pan_R*
  or *__oldPan_R*, new variables called *pan_R* and *old_pan_R* were
  defined and used. The **startUndoAction** function did, however,
  only look at the former variables, was unaware of the new
  assignments, and reverted the changes. This was fixed by moving to
  the *__* type of variable names. In addition, the panning was
  rewritten in a more readable (less obscure) way.
- A number of property changes did not feature the print of the
  applied changes as a status bar message. I added messages for all of
  them (mouse wheel and mouse click + movement)
- The status bar message was adjusted for tiny window sizes. When
  working in a tiling window manager Hydrogen, as well as its status
  bar, can get quite tiny quite quickly. To still properly see all
  changes, I moved the numerical value in front of the message
  followed by the description of the changes

What is missing is a nice way to revert the changes. This should in my
opinion be done in a batch way be reverting all applied changes at
once. But I didn't had the time to look into this matter yet.