moving NotePropertiesMode, renaming classes + NoteProperties
To make the **undoRedoAction** function working in a later point in
time, I created new structs called *NoteProperties*
and *NotePropertiesChanges*. The former one is basically a collection
of all properties and identifiers of one note (ony one state, not past
and present) used within the **undoRedoAction**. The latter is a
struct of both the old and the current property state of a **Note**
and an enumeration of type *NotePropertiesMode* to specify which
property was changed. I think its a good idea to have those structs in
the head of the **Note** class and to move to a handling of the
changes via a **get_note_properties** function. This one I have not
implemented yet! For now the undo is based on global variables and the
states provided to the **undoRedoAction** are dirty. This mean only
those properties corresponding to the NotePropertiesMode are actually
corresponding to the note. Retrieving all properties may slow down
Hydrogen. Anyway, since the **NotePropertiesMode* class was used
within the *note.h* file, I moved it there since I do not want to link
a core header against a GUI part.

In addition, I also renamed a couple of functions and classes, which I
will use in the later rewriting of the **undoRedoAction**. The new
names are corresponding more properly to their actual
functionality (some names were quite misleading).

- NotePropertiesRuler::startUndoAction ->
  NotePropertiesRuler::pushUndoAction, since it does not start an undo
  action but just pushes it onto the QUndoStack. This was quite
  confusing for me at first.
- SE_editNotePropertiesVolumeAction ->
  SE_editNotePropertiesAction, since it is used to edit all kinds of
  properties of a note; not just its velocity.
- DrumPatternEditor::undoRedoAction ->
  DrumPatternEditor::undoRedoNotePropertiesEditAction. More
  specialized to keep it apart from all other undo-redo functions.