Rewriting undo/redo in NotePropertiesRuler for multiple edits
It is now possible to revert the changes applied to properties of all
notes within a pattern by pressing Ctrl-Z.

It is also quite stunning how slow this does happens. But I save
optimization for another day.

Previously the startUndoAction function in the NotePropertiesRuler was
called without any input arguments from various mouse event
functions. It constructed a SE_editNotePropertiesAction action, which
will handle the undo/redo, using global variables set throughout all
possible event-handling function. Then, the action is pushed on the
QUndoStack. Apart from using global variables, they were also dirty,
meaning that only the ones corresponding to the note property visible
in the ruler did reflect the actual ones. All others, like pan if the
velocity was selected, were states from a previous edit or
fallbacks. This was, unfortunately, not extendable to multiple edits
and I had to rewrite some functions.

First of all the startUndoAction function was renamed in to
pushUndoAction in a previous commit and it now requires an explicit
input of a list of type NotePropertiesChanges elements. This is a
struct incorporating all the global variables used in the previous
implementation. If just a single note property is edited, the list
will contain only one element. If, instead, multiple edits did
happen, SE_editNotePropertiesAction will loop through all of them and
revert them one by one. But due to performance issues, I still grab
the dirty global variables when capturing the previous or current
state of a note upon the construction of NoteProperties.

In addition, I introduced quite a bunch of documentation.