Consistent handling of the NotePropertiesMode
Up to now there was both an enumeration H2Core::NotePropertiesModes
and a string equivalent used within the code. The former on in
e.g. the NotePropertiesRuler.cpp file specifying the current view in
the ruler and the former was used for the undo action. Now ALL
functions do use the enumerator.

In addition, a function H2Core::convertNotePropertiesModeToString was
introduced to ease the printing of the NotePropertiesMode enumerator
and the legacy **__mode** global variable in the
NotePropertiesRuler.cpp file was removed. It wasn't used by any
functions and couldn't actually be used without messing things up,
since it is one set on a mouse click event. Therefore, altering the
ruler view an changing some note properties using the mouse wheel did
not affected it.

The global variables **__oldVelocity** and **_oldProbability** were
not properly set when using the mouse wheel. This was fixed.