Consistent use of the note properties in the undo/redo action
Up to now the undo and redo action in pushUndoAction were constructed
using global variables, like *__velocity* and *__oldVelocity*, defined
within the NotePropertiesRuler. This led to the unpleasant definition
of dirty undo/redo functions. Or in other words: only those values
corresponding to the changes were the properties of the actual
notes. All others were just dirty global variables belonging to a
former state of a potentially different note. It did not really caused
any troubles but it was very fragile basis to build on.

Now, the properties of the notes are accessed using the
**Note::get_note_properties()** function before and after the action
of the user takes place. All properties in the constructed
H2Core::NotePropertiesChanges are therefore representing the actual
values. Note that the Hydrogen does not assign the __pattern_idx value
of a note properly whenever it is created. This value has to be added
manually and requires the presence of a global variable
called *__nSelectedPatternNumber*.

For grouping all actions associated with dragging one and the same
note into one undo/redo action, we still need a global state of the
initial note properties. It is stored in the global struct
NotePropertiesOld.

Some additional aspects of the code were altered:

- all occurrences of *pan_R and *pan_L were changed to *pan_r
  and *pan_l. This is how they are called in the Note object
  internally and its better to have things consistent.
- The *column*, *instrument*, and *patternNumber* enumerators of the
  H2Core::NoteProperties enumeration were renamed
  to *position*, *instrument_id*, and *pattern_idx* to be consistent
  with the internal names in the Note object was well
- The global variables containing the current state of the note
  properties, like **__leadLag** are no longer required either and
  have been removed.
- The functions **Note::get_note_properties()**
  and **Note::set_note_properties()** have been introduced to
  seemlessly use the NoteProperties struct to store all properties of
  a single note.

Performance:

I thought this consistency fix would slow down the code. But it
didn't. The undo/redo
function **DrumPattern::undoRedoNotePropertiesAction** itself did slow
down from on average 1.92e-06s to 2.99e-06s, but the overall
recalculation in the ruler is
faster. **NotePropertiesRuler::wheelEvent** sped up from 0.0335s to
0.0201s (for a simultaneous altering of 16 notes.

Known issues:

The undo/redo action works properly but the undo/redo for wheelEvents
is quite tedious. All changes applied to one or all notes have to be
reverted incrementally, Also when changing all properties using the
SHIFT+mouse movement the global state, which allows a less tedious
redo/undo is saved only for a single note. Therefore, the undo/redo is
not properly restoring the actual previous states.

There are still some time measurements using the chrono library
present. I will remove them as soon as I'm done with all the changes.