Proper undo/redo for multiple note edits
While the undo/redo for the wheelEvent will be the plain undo/redo of
the changes triggered by the user, for the mouseMoveEvent they will be
grouped together. All action between the pressing of the mouse key and
its release will be considered one action. The initial state is saved
in notePropertiesStored and assigned using storeNoteProperties and the
undo/redo action is constructed using undoMouseMovement.

But apart from the physical release of the mouse key, two events will
be also considered to be the end of one group and the onset of the
next: the pressing of Shift key and the movement of the cursor to a
different note. The latter ensures all changes applied to a single
note can be undone or redone in a single action and the user can edit
multiple individual notes without releasing the mouse button.
The __nColumnState global variable will be used to keep track of the
column the changes were applied in. The pressing or releasing of the
shift key will alter the __nShiftKeyState and allows the user to
seamlessly switch between the single note and multiple note editing
and at the same time to undo/redo all changes in a proper way.

The propertyChangesStack list, which contains all the individual edits
grouped into a single undo/redo, is now defined (and documented!) as a
global variable and cleared before usage.

Additional documentation was added and obsolete variables
__undoColumn, m_bMouseIsPressed, and __columnCheckOnXmouseMouve have
been deleted.