Document func related to audioEngine_startAudioDrivers
I added a bunch of documentation describing functions and variables
involved in the audioEngine_startAudioDrivers() function in
hydrogen.cpp.

In addition, I changed the name of the input arguments of the
setTrackOutput() function from *song* and *compo* to *pSong*
and *pCompo* to make them consistent with the naming found in other
files.