validate drumkits containing loads of instruments As encountered in issue #694 the validation of the drumkit by QXmlSchemaValidator does not work for more than 102 instruments in a drumkit.
I do not know why since 102 is obviously less than 1000, increasing the number does solve anything, and other XML Schema validators, like `xmllint`, do not show this problem. It may be a bug in the underlying Qt5 implementation.

Setting the maximum number of occurrence to `unbounded` resolves the issue.

In addition, I would advocate the change regardless of the behavior of QXmlSchemaValidator since the user is allowed to set `MAX_INSTRUMENTS` to a number bigger than 1000, which would not be noticed by the XSD file.