Fixing bug in config recreation As reported in issue #682 the rendering of H2 does not work properly if the hydrogen.default.conf file was not properly installed.
If neither the system-level config file nor the one in the user directory are not present, H2 will recreate the user config using the default parameters set in the `Preferences::Preferences()` constructor. Since the parameters `maxLayers` and `maxBars` are not set in there, they will be initialized with `0`. This causes H2 to not be properly displayed and the user has no way of changing it but to manually rewrite her config file.

In addition introduced some documentation and made the naming of the `m_iMaxLayers` variable consistent throughout the code base.

It is possible that other parameters, which are not present in the Preferences constructor either might cause additional bugs? Should we/I do some preventive measures and set ALL private parameters in the constructors or would this be an overkill?