Starting multiple H2 instance with OSC Beforehand it was not possible to start two instances of H2 at the same time since both of them tried to start up an OSC server using the same port. This caused the starting of the second one to result in a segmentation fault.
Now, whenever H2 is about to start up the OSC it uses the port specified in the preferences. If this however is not possible, it will let the liblo library, which provides the implementation of the OSC server, choose an available port number by itself. This way arbitrary many instances of H2 can be started.

TODO:
- chosen port number is not stored in Preferences even Preferences::setOscServerPort() was called.
- adding documentation for the OSC server part of the code