Warning popup on mismatching sample rate playback track Whenever a playback track is added with a sample rate, which does not match the one of the H2 engine, a warning pop up will be shown with the follwing content:
"Sample rate of the loaded track does not match the one of the audio driver!\n\nUnfortunately, Hydrogen does not support converting the sample rate itself. Please consider using tools like SoX instead."

In principle this can be enabled for all samples the user manually loads into H2. But for short ones as for drumkits this might be an overkill. Or maybe not. I also thought about setting up a popup warning window, which give the user the option to suppress all future popup windows regarding the sample rate using a boolean variable stored in the preferences. Any thoughts regarding this matter?

For now the warning message is generic and does not include the sample rate of the loaded file since the Event class only uses numeric data to transport information. If there is a need for propagating text messages to the GUI in other parts of the code, it might be a good idea to have addition support for QStrings in the Event.

I added an EventType called EVENT_MISMATCHING_SAMPLE_RATE and the corresponding handler unmatchingSampleRateEvent(). I also updated the QFileDialog window used to select files for the playback track to a) show detailed information about the files and b) only display the files supported by libsndfile. This will prevent users to load e.g. mp3 files and wonder why it does not work.

I also changed the type of EventQueue::__write_index and EventQueue::__read_index from `int` to `unsigned int` because this better reflects their intended use. And, finally, I added a bunch of documentation.