no reload of hydrogen.conf in PreferencesDialog Using the function `Preferences::loadPreferences()` the `Preferences` object will be reloaded from file.
In `PreferencesDialog::PreferencesDialog()` the settings of the user will be reloaded from her ~/.hydrogen/hydrogen.conf file before setting up the PreferencesDialog. But this will discard all the changes to the `Preferences` singleton, which are not saved yet, like the port of the OSC server.

The only reason I could think of why this behavior would be desired is that it ensures consistency between the values shown and the ones in the config file. But I think overwriting the current state with a historic one is wrong approach of ensuring this.

A way more convenient (and also a more timely one) would be to store the changes in the Preferences as soon as the `PreferencesDialog` gets closed by the user. I am actually surprised myself I have to press ctrl-s to store the changes. But I really like the fact that the dialog is created anew each time it is called. I implemented the automatic save of `Preferences` in the destructor of `PreferencesDialog`.

Apart from the general benefits I think this setting is most convenient for the OSC server itself: If the port is occupied and the user does not care about the actual number, she does not have a look at the PreferencesDialog and the new number will not be changed (if not done expicitly). But if she takes a look at it, the H2 will use the same port number she starts it up again.