adding documentation and tweaks to OSC server I added a bunch of doumentation to the OscServer class and its members.
In addition, removed the argument from the Constructor and `create_instance()` function as well as the `m_pPreferences` member. The former is used to pass a fresh copy of the `Preferences` instance and the latter is used to access `Preferencs::m_nOscServerPort`. Both can be done more readable and consistent with the overall code base by using direct calls to `Preferences::get_instance()`.

I also remove some member functions which were neither implemented nor used
- `BPM_CC_RELATIVE_Handler()`
- `BPM_FINE_CC_RELATIVE_Handler()`
- `SELECT_NEXT_PATTERN_CC_ABSOLUTE_Handler()`
- `SELECT_NEXT_PATTERN_RELATIVE_Handler()`
as well as the unused `OscClientInfo` class.