explicitly defining class name in header Unfortunately I could not get DOxygen work properly with the `H2_OBJECT` macro (although it gets expanded properly). This results in as many warning messages as classes in the code derived from the `Object` class (quite a number).
To deal with the issue, I explicitly introduced the class name `m_sClassName` and its public getter `className()` to every class inheriting from the `Object` class. This is of course way more verbose and really good style. But the warning messages were quite annoying and were covering an actual bug in the header of `Sample`.

Since both the private class name and its public getter are inherit from `Object`, it would be enough to only define the variables explicitly for all classes having static member functions and to directly hand over the name of the class when creating the object. But I would vote against it since it makes the code less consistent.