making state flag of GUI integer
since the GUI is not fully ready yet when NSM is already about to load a song, a more fine-grained handling than whether the GUI is present or not is not enough.

Instead, a value of zero now indicates there is no GUI present, smaller than zero that it is present but not ready, and larger than zero that it is both present and ready.