make NSM client switchabel
to allow non-session-manager to switch between different Songs, the restarting of the audio driver has to take place after the switching of the Song.

Thus, if the loading of the Song is done by the GUI, it has to restart the driver too. Having those two in different threads would lead to occassionaly segmentation vaults.