prevent track output reset in JackAudioDriver::connect
The `track_output_port_L/R` objects are properly reset in `JackAudioDriver::init()` which is called immediately prior to connect(). Also conceptionally there is no point in wiping all pointers to the established per-track output ports at the beginning of the connect() function.

This caused connections of makeTrackOutputs() established in JackAudioDriver::init() to be unusable.