start audio driver in main after core init
under session management the callback function of the NSM client will take care of setting the first Song and starting the audio driver. If for some reason this did not worked out, the main function will check right after the creation of the NSM client - the callback has already been called - whether the audio driver was initialized. If not, it will do so.