implement MainForm::updatePreferencesEvent + savePreferences
the MainForm does now properly handle the update of the configuration through the core part of H2. During saving, it stores the state of the GUI using savePreferences and during loading it updates some choices in the widgets

the stored of the GUI state into the Preferences instance in the closeAll function was splitted off into the savePreferences function.