check whether externally release from timebase master
Hydrogen does now check whether it is still JACK timebase master or if another application has taken over. The change in state is also propagated to the "J.MASTER" button in the GUI, which will be updated automatically. Due to the design of the JACK transport, checking for a different timebase master does only if the transport is rolling.

I also updated some comments.