jack: fix JACK transport
if a relocation was done by any JACK client, the relocation will be done _prior_ to the check for a change in BPM depending on the current transport position.

The BPM set in the timeline will only be respected if there is not JACK timebase master or Hydrogen is the timebase master itself. Else, only the tempo propagated by the timebase master does matter.