jack: remove locateInNCycles
The function was introduced in 2008 to ensure H2 will be kept in sync with Ardour.

after the Song ends, the JACK transport position will be relocated to frame 0. What this function does in addition: In the next cycle it will relocate the transport position _again_ at 0 (where it already is, so nothing will happen).

This is most probably a fix for a problem long gone. I will remove it as it just bloats the code