fix tempo change handling
The JACK timebase callback function does now broadcast the BPM the timebase master will adjust to during the next cycle. This way all the timebase client has time to adjust to the speed and both do set them together and still be in sync.

To be honest I do not know if this is the intended default behavior. There is nothing regarding this in the API. But at least two Hydrogen instances are able to work with eachother smoothly. Worst case, one still can init and disable the timebase master so none is present anymore.