drop bbt_offset support
this value specifies the amount of frames the current bar, beat, tick information is away from the beginning of the next process cycle. Up to now it was erronously set to 0 all the time. But calculating it will required a number of additional variables/complexity to cope with the changing tick size. Right now I do not see that this is worth the effort (since it was not properly implemented for the past ten years and no one did complain).