fix relocation in JACK client if transport is stopped
in the presence of an external timebase master the handling of an relocation (both master and normal client side) as off in case the transport was not running. This is now fixed by keeping track of the latest transport position using the m_currentPos variable.

In addition the body of the JackAudioDriver::updateTransportInfo function will only be called if the JACK transport is either rolling or starting.