helper functions to check for JACK transport
right now there are several places in the source code at which the usage of the JACK audio driver is queried by preprocessor directives and checking the class_name of the driver. This is, unfortunately, only half of the story because JACK transport could be deactivated via the GUI while still using JACK for sending audio to the sound card.

I introduced two helper functions handling this task and - as a side effect - make the code more readible.

The second one will check if there is an external JACK timebase master since this will cause some of the GUI elements to behave differently