proper initialization of tick size
Previously the tick size in `m_transport` of the audio driver was first set when the playback of the audio engine started. Now, it will be set as soon as a Song is available (since it requires a current BPM).

I keep the sanity check in calculateElapsedTime for a inproper initialization just to be save (in terms of other contributions/later changes).

The obsolete updateTickSize() global function was removed.