Avoid race condition on FileOpen
MainForm handles opening files via an eventFilter on the
QApplication instance, but MainForm can take a while to
initialise and be ready.  The only thing guaranteed to
receive a FileOpenEvent is the QApplication itself. So this
change subclasses QApplication and implements an event()
method that stores the requested filename so that it can be
later delivered to MainForm when it's ready.