Fix keyboard bug
Using MainForm's event filter on the QApplication instance causes
it to swallow keyboard input for keys it knows about, for all
windows, including dialogs. This structure was partially chose to
avoid subclassing QApplication for the QApplication instance, but
now that turns out to be necessary anyway, it's cleaner to instead
receive the QFileOpenEvent in the QApplication and forward that to
the MainForm explicitly.