Add timed waits for audioengine lock ( part fix for #857 )
This avoids cases where playback might be disrupted due to the GUI
grabbing the lock to update song position or edit patterns / song
structure while playing. Such locks are very short-lived, much
shorter than is typically needed to process a buffer of song data,
so waiting until the lock can be acquired is a very short delay.

To avoid excessive waits which would cause a buffer to be dropped
anyway or otherwise delivered to the sound driver too late, the
available slack time is calculated and lock waits are limited to
that slack time.

This change also changes the audioengine lock from a pthreads mutex
to a C++11 std::timed_mutex: this is because pthreads mutexes support
timed locking on Linux, FreeBSD and Windows, but bizarrely *not*
on macOS, presumably because it would make the mutex structure
larger. std::timed_mutex already implements this on all platforms,
so use it in preference.