Fix pattern editor undo issues, #909
Pattern editor undo actions relied on the current state of the
pattern to tell whether an operation is a deletion or an insertion.
This fails to get the right answer when there are multiple notes
placed via the piano roll editor (fools a drum pattern editor 'undo'
into thinking an 'undo' is to undo an insert) or a drum pattern
editor note to confuse a piano roll editor's placement of a note
on C4.

This change fixes the behaviour by storing whether the operation
to be done/undone/redone is an insertion or deletion with the undo
action.