Ctrl + drag to copy notes. Set mouse cursor to show action.
Copying notes will not behave correctly at the moment when copying to
some location that intersects another existing note, because the
action for adding a note will instead delete an existing note if it
finds it in the same position.

This will be fixed on merging #910 as that change will correct this
behaviour.