Update piano roll and note properties editors lazily
This helps immensely with batch user operations (delete, move,
copy), where there was previously a small but noticable lag on
moving notes around on a low powered machine. This was caused by
the background pixmaps for the piano roll editor and *all* the note
properties rulers being recalculated for *every* note inserted,
deleted or moved. Even when the piano roll editor is not visible,
and though only one note properties ruler is ever visible.

Making the update lazy, and only called when needed on repaint,
means this is reduced to only once per batch interaction, and in
all other cases won't bother rebuilding the piano roll editor unless
it's visible, and the hidden note properties rulers won't be rebuilt.